refactor: Move decorators.ts where it makes more sense Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>
