+++ typescript/vrotsc/src/compiler/decorators.ts
