+++ typescript/vrotsc/src/compiler/transformer/fileTransformers/workflow/decorators.ts
@@ -14,7 +14,7 @@
  */
 import * as ts from "typescript";
 import { getDecoratorNames, getIdentifierTextOrNull, getPropertyName } from "../../helpers/node";
-import { WorkflowDescriptor, WorkflowItemDescriptor, WorkflowItemType, WorkflowParameter, WorkflowParameterType } from "../../../../decorators";
+import { WorkflowDescriptor, WorkflowItemDescriptor, WorkflowItemType, WorkflowParameter, WorkflowParameterType } from "../../../decorators";
 import { getVroType } from "../../helpers/vro";
 import { DiagnosticCategory, FileTransformationContext } from "../../../../types";
