+++ typescript/vrotsc/src/compiler/transformer/fileTransformers/workflow/polyglot.ts
@@ -1,4 +1,4 @@
-import { PolyglotDescriptor } from "../../../../decorators";
+import { PolyglotDescriptor } from "../../../decorators";
 import { getIdentifierTextOrNull, getPropertyName } from "../../helpers/node";
 
 import * as ts from "typescript";
@@ -38,20 +38,20 @@ export function buildPolyglotInfo(polyglotInfo: PolyglotDescriptor, decoratorCal
 				case "package":
 					polyglotInfo.package = (<ts.StringLiteral>property.initializer).text;
 
-/*-
- * #%L
- * vrotsc
- * %%
- * Copyright (C) 2023 - 2024 VMware
- * %%
- * Build Tools for VMware Aria
- * Copyright 2023 VMware, Inc.
- * 
- * This product is licensed to you under the BSD-2 license (the "License"). You may not use this product except in compliance with the BSD-2 License.
- * 
- * This product may include a number of subcomponents with separate copyright notices and license terms. Your use of these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE file.
- * #L%
- */
+					/*-
+					 * #%L
+					 * vrotsc
+					 * %%
+					 * Copyright (C) 2023 - 2024 VMware
+					 * %%
+					 * Build Tools for VMware Aria
+					 * Copyright 2023 VMware, Inc.
+					 *
+					 * This product is licensed to you under the BSD-2 license (the "License"). You may not use this product except in compliance with the BSD-2 License.
+					 *
+					 * This product may include a number of subcomponents with separate copyright notices and license terms. Your use of these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE file.
+					 * #L%
+					 */
 					break;
 				case "method":
 					polyglotInfo.method = (<ts.StringLiteral>property.initializer).text;
